+++ python/utils/generate_dictionary.py
@@ -1,0 +1,29 @@
+import itertools
+import string
+from abc import ABC, abstractmethod
+import sys
+
+def generate_dictionary():
+    """
+    The dictionary is generated dynamically due to the following reasons:
+    - Maintainability: Easier to maintain if the Braille representations or English characters change.
+    - Efficiency: Generally efficient because it uses pre-computed indices and is based on a structured approach to generate the dictionary.
+    - Readability: The code might be more complex but provides a clear, systematic way to generate the dictionary.
+    """
+    
+    states = ['O', '.'] # Possible states for each dot: colored (O) or not colored (.)
+    combinations = list(itertools.product(states, repeat=6)) # Because its 2 possibilities for each dot and 6 dots in total = 2 ^ 6 = 64 combinations
+    braille = list(''.join(combo) for combo in combinations) # Convert the tuples to strings
+    english = list(string.ascii_lowercase + string.digits + '.'+ ','+'?'+'!'+':'+';'+'-'+'/'+'<'+'('+')'+' '+'C'+'D'+'N') # Create a list of all lowercase letters, digits, and punctuation
+
+    # Braille indices for each English character in the order of the English list
+    braille_indices=[31,23,15,11,27,7,3,19,39,35,29,21,13,0,25,5,1,17,37,33,28,20,34,12,8,24,35,31,23,15,11,27,7,3,19,39,50,55,52,49,51,53,60,45,38,22,41,63,62,46,40]
+    if len(english) != len(braille_indices):
+        raise ValueError("Mismatch between English characters and Braille indices length.")
+
+    # Time Complexity: O(n) where 'n' is the length of 'english' list
+    translate_dict = {}
+    for index, character in enumerate(english):
+        translate_dict[character] = braille[braille_indices[index]]
+        
+    return translate_dict
